moved templates in ui folder
